# [EUWE] Further improve the /api/service endpoint