--- a/app/models/service.rb
+++ b/app/models/service.rb
@@ -39,7 +39,6 @@ class Service < ApplicationRecord
   virtual_has_many   :vms
   virtual_has_many   :all_vms
   virtual_has_many   :power_states, :uses => :all_vms
-  virtual_total      :v_total_vms, :vms
 
   virtual_has_one    :custom_actions
   virtual_has_one    :custom_action_buttons, :uses => {:service_template => :custom_action_buttons}
@@ -63,6 +62,9 @@ class Service < ApplicationRecord
 
   include_concern 'RetirementManagement'
   include_concern 'Aggregation'
+
+  virtual_total :v_total_vms, :vms,
+                :arel => aggregate_hardware_arel("v_total_vms", vms_tbl[:id].count, :skip_hardware => true)
 
   virtual_column :has_parent,                               :type => :boolean
   virtual_column :power_state,                              :type => :string