@abellotti 2017-11-14
Would be great to have on master too, I think API side could be generalized and more fine-tuned to support attributes= list specified by user, I can play with that a bit.  We could use the Model side (v_total_vms change) PR on master if not there already.