@NickLaMuro 2017-11-14
> We could use the Model side (v_total_vms change) PR on master if not there already.

@abellotti yeah, for sure.  I will try and put together a PR for that today yet.  Possibly remove that code from this PR so we can backport that separately, and just do the controller changes in this PR.