@Fryguy 2017-11-14
>  I will try and put together a PR for that today yet.

I still haven't made the corresponding PR this is based on.  The code is quite different on master.