@NickLaMuro 2017-11-14
@Fryguy Are you looking at the wrong PR?  I was referring to the model code changes I made here that I would be making the PR for, not the controller.  Those changes should be almost no different, and can be applied in the same fashion for both versions of the code base (hopefully).