@NickLaMuro 2017-11-14
To further clarify, I would create a PR with just the first 2 commits of this PR on master.  We can then backport that to EUWE and G, and then create separate PRs for EUWE and G for implementing the controller functionality.