@NickLaMuro 2017-11-14
Master port of the model code:  https://github.com/ManageIQ/manageiq/pull/16476